requirements.txt
pandas
numpy
scikit-learn
imbalanced-learn
xgboost
pyyaml
joblib
matplotlib
seaborn
pytest
